- Added secrets into the framework
For the Lab-as-a-Service pattern to work in a similar fashion to how it
does today we are loading global secrets into the deployment namespace
so that we can reuse them (through externalsecrets operator).

All of the deployments to meet initial LaaS requirements will come from
the hive namespace. Eventually, we will support custom namespaces with
custom secrets like we have in the initial MVP.

The secrets template has been updated with where the pattern is
expecting configurations to be on the local filesystem